salon chain crud finished

+++ src/app/theme/pages/default/components/chains/edit-chain/edit-chain.component.ts
@@ -21,7 +21,6 @@ export class EditChainComponent implements OnInit, AfterViewInit {
     // chain: any;
     chain: any = {
         title: '',
-        price: {level: ''},
         description: '',
         phone_number: '',
         levels: []
@@ -42,17 +41,18 @@ export class EditChainComponent implements OnInit, AfterViewInit {
     }
 
     removePriceLevel(index: number) {
-        this.chain.levels.splice(index, 1)
+        this.chain.levels.splice(index, 1);
     }
 
     getChain() {
         this.chainService.getChain(this.chainId)
             .subscribe(
                 (response) => {
-                    console.log(response.data.chain);
+                    console.log(response);
                     this.chain = response.data.chain;
-                    this.chain.price = response.data.chain.levels[response.data.chain.levels.length - 1];
+                   /* this.chain.price = response.data.chain.levels[response.data.chain.levels.length - 1];
                     this.chain.levels.splice(-1, 1);
+                    */
                     this.ph = response.data.chain.phone_number;
                     parseInt(this.ph.slice(1, 2) + this.ph.slice(4, 7) + this.ph.slice(9, 12) + this.ph.slice(13, 15) + this.ph.slice(16, 17));
                     this.phone_number.nativeElement.value = this.ph
@@ -68,7 +68,6 @@ export class EditChainComponent implements OnInit, AfterViewInit {
                 }
             );
         this.getChain();
-        console.log(this.currentUser.chain.id);
         if(this.currentUser.chain.id == this.chainId) {
             this.deleteButtonShow = false;
         }
@@ -81,22 +80,23 @@ export class EditChainComponent implements OnInit, AfterViewInit {
 
     onSubmit() {
         let n = -1;
-        // console.log(this.signupForm.value.price);
-        this.chain.price.level = this.signupForm.value.price;
+        // console.log(this.signupForm.value);
+        // this.chain.price.level = this.signupForm.value.price;
         this.chain.title = this.signupForm.value.title;
         this.chain.description = this.signupForm.value.description;
         this.chain.phone_number = this.phone_number.nativeElement.value;
         for (let i in this.signupForm.value.levels) {
+            console.log(10);
             n++;
             this.chain.levels[n].level = this.signupForm.value.levels['price' + n];
             // console.log(this.signupForm.value.levels['price' + n]);
             // console.log(this.chain.levels[n].level);
         }
-        this.chain.levels.push(this.chain.price);
+        // this.chain.levels.push(this.chain.price);
+
         this.chainService.editChain(this.chain)
             .subscribe(
                 (response) => {
-                    console.log(response);
                     if (response.data.status == "OK") {
                         this.router.navigate(['/components/chains'], {relativeTo: this.route});
                     }
@@ -105,15 +105,29 @@ export class EditChainComponent implements OnInit, AfterViewInit {
     }
 
     deleteChain() {
-        this.chainService.deleteChain(this.chain.id)
-            .subscribe(
-                (response) => {
-                    console.log(response);
-                    if (response.success == 1) {
-                        this.router.navigate(['/components/chains'], {relativeTo: this.route});
+        if (this.chain.salonsCount > 0) {
+            let result = confirm("В этом сети салонов есть салоны вы уверены что хотите удалить сеть?");
+            if(result) {
+                this.chainService.deleteChain(this.chain.id)
+                    .subscribe(
+                        (response) => {
+                            if (response.success == 1) {
+                                this.router.navigate(['/components/chains'], {relativeTo: this.route});
+                            }
+                        }
+                    );
+            }
+        } else {
+            this.chainService.deleteChain(this.chain.id)
+                .subscribe(
+                    (response) => {
+                        if (response.success == 1) {
+                            this.router.navigate(['/components/chains'], {relativeTo: this.route});
+                        }
                     }
-                }
-            );
+                );
+        }
+
     }
 
     ngAfterViewInit() {
